Extract more shared code and refine handlers
