+++ backend/ondisconnect/src/on-disconnect.ts
@@ -1,42 +1,9 @@
-import { Config } from './shared/backendTypes';
+import { Config } from './sharedBackend/config';
+import { removeConnection } from './sharedBackend/removeConnection';
 
-export const onDisconnect = async ({ ddb, connectionId, tableName }: Config) => {
+export const onDisconnect = async (config: Config) => {
   try {
-    // TODO Lukas this could be shared code?
-    // TODO Lukas can we run requests together?
-    const connectionItem = (
-      await ddb
-        .get({
-          TableName: tableName,
-          ConsistentRead: true,
-          Key: {
-            primaryKey: `connectionId:${connectionId}`,
-          },
-        })
-        .promise()
-    ).Item;
-    await ddb
-      .delete({
-        TableName: tableName,
-        Key: {
-          primaryKey: `connectionId:${connectionId}`,
-        },
-      })
-      .promise();
-    if (connectionItem && connectionItem.groupId) {
-      await ddb
-        .update({
-          TableName: tableName,
-          Key: {
-            primaryKey: `groupId:${connectionItem.groupId}`,
-          },
-          UpdateExpression: 'REMOVE connections.#1.connectionId',
-          ExpressionAttributeNames: {
-            '#1': connectionItem.userId,
-          },
-        })
-        .promise();
-    }
+    await removeConnection(config);
   } catch (err) {
     return { statusCode: 500, body: 'Failed to disconnect: ' + JSON.stringify(err) };
   }
