+++ backend/sendmessage/src/reset-votes.ts
@@ -1,22 +1,35 @@
-import { GroupItem } from './types';
-import { getConnectionItem, getGroupItem } from './get-item';
 import { broadcastState } from './broadcast-state';
-import { VOTE_NOTE_VOTED } from './shared/WebSocketMessages';
-import { Config, ConfigWithHandler } from './shared/backendTypes';
+import { VOTE_NOTE_VOTED } from './shared/cards';
+import { Config, ConfigWithHandler } from './sharedBackend/config';
+import { getConnectionItem } from './sharedBackend/getConnectionItem';
+import { getGroupItem, GroupConnections } from './sharedBackend/getGroupItem';
 
-export function persistResetVotes(
-  groupItem: GroupItem,
+export const resetVotes = async (config: ConfigWithHandler) => {
+  const connectionItem = await getConnectionItem(config);
+  if (!connectionItem) return;
+  const { groupId } = connectionItem;
+  if (!groupId) return;
+  const groupItem = await getGroupItem(groupId, config);
+  if (!groupItem) return;
+  await resetPersistedVotes(groupId, groupItem.connections, false, config);
+  await broadcastState(groupId, config);
+};
+
+export const resetPersistedVotes = (
   groupId: string,
+  connections: GroupConnections,
+  updateScale: string[] | false,
   { tableName, ddb }: Config
-): Promise<unknown> {
-  const userIds = Object.keys(groupItem.connections).filter(
-    (userId) => groupItem.connections[userId].vote !== 'observer'
+): Promise<unknown> => {
+  const userIds = Object.keys(connections).filter(
+    (userId) => connections[userId].vote !== 'observer'
   );
   const userIdAttributeNames = Object.fromEntries(
     userIds.map((userId, index) => [`#${index}`, userId])
   );
   const updates = [
     'visible = :visible',
+    ...(updateScale ? ['scale = :scale'] : []),
     ...Object.keys(userIdAttributeNames).map(
       (attributeName) => `connections.${attributeName}.vote = :notVoted`
     ),
@@ -31,17 +44,9 @@ export function persistResetVotes(
       ExpressionAttributeValues: {
         ':visible': false,
         ':notVoted': VOTE_NOTE_VOTED,
+        ...(updateScale ? { ':scale': updateScale } : {}),
       },
       ExpressionAttributeNames: userIdAttributeNames,
     })
     .promise();
-}
-
-export async function resetVotes(config: ConfigWithHandler) {
-  const { groupId } = await getConnectionItem(config);
-  if (!groupId) return;
-  const groupItem = await getGroupItem(groupId, config);
-  if (!groupItem) return;
-  await persistResetVotes(groupItem, groupId, config);
-  await broadcastState(groupId, config);
-}
+};
