+++ backend/sendmessage/src/shared
@@ -1,1 +1,1 @@
-../../../shared
+../../../frontend/src/shared
