+++ frontend/src/Components/ResultsPage/ResultsPage.tsx
@@ -1,3 +1,6 @@
+import { COLUMN_NAME, COLUMN_VOTE, HEADING_RESULTS } from '../../constants';
+import { CardValue, VOTE_COFFEE, VOTE_NOTE_VOTED, VOTE_OBSERVER } from '../../shared/cards';
+import { Votes } from '../../shared/serverMessages';
 import sharedClasses from '../../styles.module.css';
 import { IconCoffee } from '../IconCoffee/IconCoffee';
 import { IconNotVoted } from '../IconNotVoted/IconNotVoted';
@@ -5,14 +8,6 @@ import { IconObserver } from '../IconObserver/IconObserver';
 import { connectToWebSocket } from '../WebSocket/WebSocket';
 import { compareVotes } from './compareVotes';
 import classes from './ResultsPage.module.css';
-import { COLUMN_NAME, COLUMN_VOTE, HEADING_RESULTS } from '../../constants';
-import {
-  CardValue,
-  VOTE_COFFEE,
-  VOTE_NOTE_VOTED,
-  VOTE_OBSERVER,
-  Votes,
-} from '../../shared/WebSocketMessages';
 
 const getSortedResultsArray = (unsortedResults: Votes) => {
   const dataArray: [string, CardValue][] = Object.entries(unsortedResults);
