+++ frontend/src/Components/ResultsPage/compareVotes.ts
@@ -1,4 +1,4 @@
-import { CardValue } from '../../shared/WebSocketMessages';
+import { CardValue } from '../../shared/cards';
 
 export const compareVotes = (
   userAndVote1: [string, CardValue],
