+++ frontend/src/Components/RevealButton/RevealButton.test.tsx
@@ -1,7 +1,7 @@
-import { getRenderWithWebSocket } from '../../test-helpers/renderWithWebSocket';
 import { fireEvent } from '@testing-library/preact';
+import { SCALES } from '../../shared/scales';
+import { getRenderWithWebSocket } from '../../test-helpers/renderWithWebSocket';
 import { RevealButton } from './RevealButton';
-import { SCALES } from '../../constants';
 
 const render = getRenderWithWebSocket(<RevealButton />, {
   connected: true,
