--- frontend/src/shared
@@ -1,1 +1,0 @@
-../../shared
