+++ frontend/tsconfig.json
@@ -15,7 +15,7 @@
     "sourceMap": true,
     "strict": true,
     "target": "ESNext",
-    "types": ["vite/client"]
+    "types": ["vite/client", "node"]
   },
   "include": ["src/**/*"],
   "exclude": ["src/**/*.test.*"]
